:memo: Documents Slack controller
Added documentation for Slack controller endpoints. Abstracted middleware to separate slackMiddleware file and _openChannel helper to slackHelpers file since it is used in multiple files. Adjusted model query language in middleware. Deleted unused and unnecessary Slack endpoints, added clarification for GET /api/slack/:id/history and POST /api/slack/sendMessageNow since they are currently either unused or for testing purposes only, but too critical to future app goals to remove.
